pyproject.toml
[project]
name = "sports-results-agent"
version = "0.1.0"
description = "Multi-agent system with Azure AI Agents integration for routing and task delegation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.2.1",
    "openai-agents>=0.1.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "a2a-sdk>=0.3.5",
]
